Clean up test file
